pawn replacement chosing added
There is a new div with the thing that allows the user to chose and you can add ids and classes to the buttons depending on what you want to do with it. Also there is a function at the bottom of the javascript page with the function that makes the box appear.